requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
python-dotenv
notebook
ipykernel
langchain-groq
fastapi
uvicorn
